Very basic structure for multiple connections
